Move some reader error checking to cursor creation
This approach is an attempt to front-load as much error checking as
possible to when we create a cursor.  Doing this allows us to avoid
error checking later when we're using the cursor, which I feel is more
error-prone overall, since it's harder to keep track of which
preconditions we have and haven't already checked for.